requirements.txt
langchain-openai~=0.3.18
langchain~=0.3.25
langgraph~=0.4.7
python-dotenv~=1.1.0